2.19

+++ mongodb/utils.js
@@ -1,21 +1,62 @@
 /* eslint-disable */
+const express = require('express')
 const path = require('path')
 const fs = require('fs')
+const multer = require('multer')
+const FfmpegCommand = require('fluent-ffmpeg')
+const ffmpegPath = require('@ffmpeg-installer/ffmpeg').path
+FfmpegCommand.setFfmpegPath(ffmpegPath)
 
 // 图片路径
 const dyimgPath = path.join(__dirname, '../dyimg')
 // 视频路径
 const dyvideoPath = path.join(__dirname, '../dyvideo')
-// 文件删除
-const deletFile = (path) => {
+// 查找文件
+const findFile = path => {
+  return fs.existsSync(path)
+}
+// 读取文件
+const readFile = path => {
   return new Promise((resolve, reject) => {
-    fs.unlink(path, err => {
+    fs.readFile(path, (err, data) => {
       if (err) {
         reject(err)
       } else {
-        resolve(true)
+        resolve(data)
       }
     })
+  })
+}
+// 删除文件
+const deletFile = path => {
+  return new Promise((resolve, reject) => {
+    if (fs.existsSync(path)) {
+      fs.unlink(path, err => {
+        if (err) {
+          reject(err)
+        } else {
+          resolve(true)
+        }
+      })
+    } else {
+      resolve(true)
+    }
+  })
+}
+// 重命名文件
+const renameFile = (path, newPath) => {
+  return new Promise((resolve, reject) => {
+    if (fs.existsSync(path)) {
+      fs.rename(path, newPath, err => {
+        if (err) {
+          reject(err)
+        } else {
+          resolve(true)
+        }
+      })
+    } else {
+      reject('未找到文件!')
+    }
   })
 }
 
@@ -76,10 +117,16 @@ function create (param, schema) {
 }
 
 module.exports = {
+  express,
   path,
+  multer,
+  FfmpegCommand,
   dyimgPath,
   dyvideoPath,
+  findFile,
+  readFile,
   deletFile,
+  renameFile,
   checkParam,
   findOne,
   checkExist,
